tesis/servidor/requirements.txt
paho-mqtt-1.1
Flask-PyMongo
